Fixed an JSON parsing issue.
If strings are passed in HTML that contains spaces, the browser will interpret this as the end of a string, which breaks JSON parsing.
Now, if the data gets parsed to JSON, all spaces are replaced with "_". Of course, this isn't for the text that gets presented to the user.